fix: issues with release to pypi (#228)
1. The builds in `ubuntu` failed with the error: 

> The system library `openssl` required by crate `openssl-sys` was not found.

2. The builds for both `windows` and `macos` passed, but the workflows failed  with the error: 

> Failed request: (409) Conflict: an artifact with this name already exists on the workflow run

**Why?**

The error for `ubuntu` was due to the library `hf-hub,` specified as a Rust dependency, but that was not being used.

The errors in the builds of `windows` and `macos` were due to breaking changes in version v4 of the GitHub action upload artifact—more details are in [here](https://github.com/actions/upload-artifact/issues/478). 

**Solution**

The library was removed, and the steps were modified; both changes are included in this PR.

**How can we prevent this from happening in the future?**

To prevent such failures in the future in the middle of the actual release, a pre-release workflow was created (and included in this PR). The idea is that this pre-release will try to upload to TestPyPI, and in case of any failures, we will modify both `.yaml` files (`ci-publish.yaml` and `ci-pre-release.yaml`)

**Additionally**
Refactor out ibis.datafusion/postgres/to_sql/pandas