chore: add workflow for checking PR titles (#716)
The npm library commitlint is proposed for linting the commit message.
The message is linted following the rules in the file
`.commitlintrc.json`.
Rules are made up of a name and a configuration array. The configuration
array contains:
- Level [0..2]: 0 disables the rule. The 1 stands for warning, and 2 for
an error.
- Applicable always|never: never inverts the rule.
 - Value: value to use for this rule.

For example:

```
"body-max-line-length": [2, "always", 1000]
```
This means the maximum number of lines in the message's body is 1000;
otherwise, it will err.
The rule is always applied.

Another example:

```
    "subject-case": [
      2,
      "never",
      ["sentence-case", "start-case", "pascal-case", "upper-case"]
    ]
```

Means the case of the subject must never be one of:

```
"upper-case", // UPPERCASE
"pascal-case", // PascalCase
"sentence-case", // Sentence case
"start-case", // Start Case
```

More info can be found in:
- https://commitlint.js.org/reference/rules.html
- https://commitlint.js.org/reference/rules-configuration.html