chore(rust): migrate to xorq-datafusion (#792)
With this change, we migrate the Rust functionality to its repository.
This has the following benefits:
- Faster CI times, no need to build the Rust side every time
- Less cache space used (no need to cache Rust artifacts)
- Faster publishing times, most of the time was dedicated to building
the wheels for Windows
- More contributor friendly (no need to have Rust installed or knowledge
of it)
- Automated updates via renovatebot (currently broken because Rust is
missing in the environment,
[example](https://github.com/xorq-labs/xorq/pull/791#issuecomment-2769911324))

Since removing all Rust code, we need a new [build
backend](https://peps.python.org/pep-0517/#source-trees). The most
popular option seems to be
[hatchling](https://hatch.pypa.io/1.12/config/build/#build-system),

```
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
```

Mainly because:

> Hatchling is a standards-compliant
[1](https://hatch.pypa.io/1.13/config/build/#fn:1) build backend and is
a dependency of Hatch itself.

More on why hatch can be found
[here](https://hatch.pypa.io/1.12/why/#build-backend)

To ensure that `uv build` does not produce empty wheels, we need to add
the following configuration
to `pyproject.toml`:

```
[tool.hatch.build.targets.wheel]
sources = ["python"]

[tool.hatch.build.targets.sdist]
only-include = ["python"]
```

For more details, see this
[issue](https://github.com/astral-sh/uv/issues/10066)

**Alternatives**

A possible alternative is to use [uv
workspaces](https://docs.astral.sh/uv/concepts/projects/workspaces/),
which will keep all the code in a single repo but has the following
drawbacks:

- renovatebot will continue to be broken
- The repo will be more cluttered (additional workflows required to
handle the release of both libraries)
- loses the functionality of publishing on release, and the release
history of both packages is going to be intertwined

closes #786, #117

---------

Co-authored-by: dlovell <dlovell@gmail.com>